ImageEditor/requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.9.24
charset-normalizer==2.1.1
cloudinary==1.30.0
colorama==0.4.5
crayons==0.4.0
deprecation==2.1.0
dj-database-url==1.0.0
Django==3.2.15
django-cloudinary-storage==0.3.0
django-crontab==0.7.1
django-environ==0.11.2
easyocr==1.7.2
filelock==3.16.1
fsspec==2024.10.0
gunicorn==20.1.0
idna==3.4
imageio==2.35.1
img2pdf==0.4.4
jinja2==3.1.4
lazy-loader==0.4
lxml==4.9.1
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
ninja==1.11.1.1
numpy==1.23.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
opencv-python==4.6.0.66
opencv-python-headless==4.10.0.84
packaging==21.3
piexif==1.1.3
pikepdf==6.2.0
Pillow==9.2.0
psycopg2==2.9.4
pyclipper==1.3.0.post6
pycodestyle==2.9.1
pyparsing==3.0.9
pypng==0.20220715.0
python-bidi==0.6.3
python-decouple==3.8
python-dotenv==0.21.0
pytz==2022.2.1
PyWavelets==1.4.1
PyYAML==6.0.2
qrcode==7.4.2
requests==2.28.1
scikit-image==0.21.0
scipy==1.10.1
shapely==2.0.6
six==1.16.0
sqlparse==0.4.3
stegano==0.10.2
sympy==1.13.3
tifffile==2023.7.10
toml==0.10.2
torch==2.4.1
torchvision==0.19.1
triton==3.0.0
typing-extensions==4.12.2
urllib3==1.26.12
whitenoise==6.2.0
